Rename UsersYamlSwapper
