+++ shardingsphere-infra/shardingsphere-infra-common/src/test/java/org/apache/shardingsphere/infra/metadata/user/GranteeTest.java
@@ -15,18 +15,17 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.infra.auth.builtin.model.user;
+package org.apache.shardingsphere.infra.metadata.user;
 
-import org.apache.shardingsphere.infra.metadata.user.Grantee;
 import org.junit.Test;
 
 import static org.hamcrest.CoreMatchers.is;
-import static org.junit.Assert.assertFalse;
-import static org.junit.Assert.assertTrue;
+import static org.hamcrest.CoreMatchers.not;
 import static org.junit.Assert.assertThat;
+import static org.junit.Assert.assertTrue;
 
 public final class GranteeTest {
-
+    
     @Test
     public void assertEquals() {
         Grantee grantee = new Grantee("name", "%");
@@ -35,13 +34,13 @@ public final class GranteeTest {
         assertTrue(grantee.equals(grantee1));
         assertTrue(grantee.equals(grantee2));
     }
-
+    
     @Test
     public void assertHashcode() {
         Grantee grantee = new Grantee("name", "%");
         Grantee grantee1 = new Grantee("name", "");
         Grantee grantee2 = new Grantee("name", "127.0.0.1");
         assertThat(grantee.hashCode(), is(grantee1.hashCode()));
-        assertFalse(grantee1.hashCode() == grantee2.hashCode());
+        assertThat(grantee1.hashCode(), not(grantee2.hashCode()));
     }
 }
